Extracted model-related tests into 'XGBoostModelSuite'
